+++ util/random.cpp
@@ -11,7 +11,7 @@ namespace funtls { namespace util {
 void get_random_bytes(void* dest, size_t count) {
 #ifdef WIN32
 	auto d = reinterpret_cast<uint8_t*>(dest);
-	while (count--) *d++ = std::random_device()();
+	while (count--) *d++ = static_cast<uint8_t>(std::random_device()());
 #elif 1
     std::ifstream urandom("/dev/urandom", std::ifstream::binary);
     if (!urandom || !urandom.is_open()) {
